refined LoadShaderParameter function with better support for node arrays
since we are reusing shader parsing functions the operator's 'inputs' parameter was using the wrong method for node arrays, the existing code would call AiNodeLink on each array item. this was causing arnold to spit back warnings about not being able to link the nodes.
note about this implementation... the array is allocated and set prior to parsing the array values, this means we need to retrieve the array at a later time once the value of the array item is known. then we can set that value in the array and set it on the node again.